able to use shared state for same reducer class
1. store shared state in thunkReducer.NODE_STATE_MAP_BY_CLASS.
2. store shared dispatch-map in useReducer.DISPATCH_MAP_BY_CLASS.
3. pass test: module-no-default-shared.
4. pass test: module-no-default-pure-shared.
5. pass test: module-no-default-having-child.
6. clean code.
7. git submodule third-party/react-hook-thunk-reducer.
