+++ src/addChild.ts
@@ -1,0 +1,13 @@
+import { type AddRelationAction, reduceAddRelation } from './addRelation'
+import { type ClassState, type NodeMeta, Relation, type State } from './stateTypes'
+
+export const ADD_CHILD = 'react-reducer-utils/ADD_CHILD'
+export const addChild = (myID: string, child: NodeMeta): AddRelationAction => ({
+  myID,
+  type: ADD_CHILD,
+  relaton: child,
+})
+
+export const reduceAddChild = <S extends State>(state: ClassState<S>, action: AddRelationAction): ClassState<S> => {
+  return reduceAddRelation(state, action, Relation.CHILDREN)
+}
