+++ tests/module-no-default-shared.test.tsx
@@ -1,9 +1,9 @@
 import { act, useEffect, useState } from 'react'
 import ReactDOM from 'react-dom/client'
 import { afterEach, beforeEach, expect, test } from 'vitest'
-import { genUUID, getRootID, getState, type ModuleToFunc, StateType, useReducer } from '../src/index'
-import Child from './TheChild'
-import Parent from './TheParent'
+import { cleanShared, genUUID, getRootID, getState, type ModuleToFunc, useReducer } from '../src/index'
+import Child from './TheSharedChild'
+import Parent from './TheSharedParent'
 import * as DoChild from './theChild'
 import * as DoParent from './theParent'
 
@@ -29,14 +29,17 @@ afterEach(() => {
 
   document.body.removeChild(container)
   container = null
+
+  console.info('module-no-default-shared: afterEach: to CleanSharedState')
+  cleanShared()
 })
 
 type Props = {}
 
-test('children-no-default-module (init and remove)', {}, () => {
+test.sequential('module-no-default-shared (init and remove)', {}, () => {
   const App = (props: Props) => {
-    const [stateParent, doParent] = useReducer<DoParent.State, TDoParent>(DoParent, StateType.LOCAL)
-    const [stateChild, doChild] = useReducer<DoChild.State, TDoChild>(DoChild, StateType.LOCAL)
+    const [stateParent, doParent] = useReducer<DoParent.State, TDoParent>(DoParent)
+    const [stateChild, doChild] = useReducer<DoChild.State, TDoChild>(DoChild)
     const [parentID, _1] = useState(genUUID())
     const [childID1, _2] = useState(genUUID())
     const [childID2, _3] = useState(genUUID())
@@ -48,21 +51,24 @@ test('children-no-default-module (init and remove)', {}, () => {
       doChild.init(childID2, DoParent.myClass, parentID, doParent)
     }, [doParent.init, doChild.init])
 
-    const parent = getState(stateParent) || DoParent.defaultState
     const parentRootID = getRootID(stateParent)
-    const child1 = getState(stateChild, childID1) || DoChild.defaultState
-    const child2 = getState(stateChild, childID2) || DoChild.defaultState
+
+    console.info('module-no-default-shared: to render: parentRootID:', parentRootID, 'parentID:', parentID)
+
+    const state = getState(stateParent, parentID) || DoParent.defaultState
 
     return (
       <>
         <span className='parent-root-id'>{parentRootID}</span>
         <span className='parent-id'>{parentID}</span>
-        <Parent myID={parentID} state={parent} do={doParent} />
-        <Child myID={childID1} state={child1} do={doChild} />
-        <Child myID={childID2} state={child2} do={doChild} />
+        <Parent myID={parentID} state={state} do={doParent} />
+        <Child myID={childID1} />
+        <Child myID={childID2} />
       </>
     )
   }
+
+  console.info('module-no-default-shared: to act')
 
   // do act
   act(() => {
@@ -80,7 +86,7 @@ test('children-no-default-module (init and remove)', {}, () => {
 
   const parentRootIDs = container.getElementsByClassName('parent-root-id')
   const parentIDs = container.getElementsByClassName('parent-id')
-  const parentLocalRootIDs = container.getElementsByClassName('parent-local-root-id')
+
   const parentIsSames = container.getElementsByClassName('parent-is-same')
 
   expect(parentRootIDs.length).toBe(1)
@@ -89,10 +95,8 @@ test('children-no-default-module (init and remove)', {}, () => {
   expect(parentButtons.length).toBe(1)
   expect(childDivs.length).toBe(2)
   expect(childButtons.length).toBe(2)
-  expect(parentLocalRootIDs.length).toBe(1)
   expect(parentIsSames.length).toBe(1)
 
-  expect(parentLocalRootIDs[0].textContent).not.toBe(parentRootIDs[0].textContent)
   expect(parentRootIDs[0].textContent).toBe(parentIDs[0].textContent)
   expect(parentDivs[0].textContent).toBe('0')
   expect(childDivs[0].textContent).toBe('0')
@@ -104,11 +108,11 @@ test('children-no-default-module (init and remove)', {}, () => {
   })
 
   expect(parentRootIDs[0].textContent).toBe(parentIDs[0].textContent)
-  expect(parentDivs[0].textContent).toBe('1')
+  expect(parentDivs[0].textContent).toBe('2')
   expect(childDivs[0].textContent).toBe('0')
   expect(childDivs[1].textContent).toBe('0')
 
-  expect(parentIsSames[0].textContent).toBe('false')
+  expect(parentIsSames[0].textContent).toBe('true')
 
   // click parent button (2st)
   act(() => {
@@ -116,7 +120,7 @@ test('children-no-default-module (init and remove)', {}, () => {
   })
 
   expect(parentRootIDs[0].textContent).toBe(parentIDs[0].textContent)
-  expect(parentDivs[0].textContent).toBe('2')
+  expect(parentDivs[0].textContent).toBe('4')
   expect(childDivs[0].textContent).toBe('0')
   expect(childDivs[1].textContent).toBe('0')
 
@@ -126,7 +130,7 @@ test('children-no-default-module (init and remove)', {}, () => {
   })
 
   expect(parentRootIDs[0].textContent).toBe(parentIDs[0].textContent)
-  expect(parentDivs[0].textContent).toBe('2')
+  expect(parentDivs[0].textContent).toBe('4')
   expect(childDivs[0].textContent).toBe('1')
   expect(childDivs[1].textContent).toBe('0')
 
@@ -136,7 +140,7 @@ test('children-no-default-module (init and remove)', {}, () => {
   })
 
   expect(parentRootIDs[0].textContent).toBe(parentIDs[0].textContent)
-  expect(parentDivs[0].textContent).toBe('2')
+  expect(parentDivs[0].textContent).toBe('4')
   expect(childDivs[0].textContent).toBe('2')
   expect(childDivs[1].textContent).toBe('0')
 
@@ -146,7 +150,7 @@ test('children-no-default-module (init and remove)', {}, () => {
   })
 
   expect(parentRootIDs[0].textContent).toBe(parentIDs[0].textContent)
-  expect(parentDivs[0].textContent).toBe('2')
+  expect(parentDivs[0].textContent).toBe('4')
   expect(childDivs[0].textContent).toBe('3')
   expect(childDivs[1].textContent).toBe('0')
 
@@ -156,7 +160,7 @@ test('children-no-default-module (init and remove)', {}, () => {
   })
 
   expect(parentRootIDs[0].textContent).toBe(parentIDs[0].textContent)
-  expect(parentDivs[0].textContent).toBe('2')
+  expect(parentDivs[0].textContent).toBe('4')
   expect(childDivs[0].textContent).toBe('3')
   expect(childDivs[1].textContent).toBe('1')
 })
