+++ tests/theChild.ts
@@ -19,10 +19,10 @@ export interface State extends rState {
   count: number
 }
 
-export const defaultState: State = {
+export const defaultState: State = Object.freeze({
   myID: '',
   count: 0,
-}
+})
 
 export const init = (
   myID: string,
